xfstests: btrfs/015: add a regression test for snapshot creation
We had a regression where you couldn't snapshot a file system if
you mounted it ro and then remounted it rw.  This is a test that
does just that to make sure we don't have this problem again.  I
ran the test without the fix and it blew up, and then applied the
fix and verified that it passed.

[rjohnston: renumbered test]

Signed-off-by: Josef Bacik <jbacik@fusionio.com>
Reviewed-by: Stefan Behrens <sbehrens@giantdisaster.de>
Signed-off-by: Rich Johnston <rjohnston@sgi.com>