xfstests: Test xfs_fsr / swapext management of di_forkoff w/ selinux
In some cases selinux's creation of an xattr on the temporary
fd creates a local xattr, but the file we are trying to 
defragment has attrs in extent format, and the forkoff mismatch
will cause xfs_fsr to fail.  This test demonstrates it;  I
have old patches sent to the list long ago that should fix
it.  I'll resend them soon.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Eryu Guan <eguan@redhat.com>
Signed-off-by: Rich Johnston <rjohnston@sgi.com>